// src/types.d.ts
export type ParserOption =
    | 'babylon'
    | 'flow'
    | 'postcss' // deprecated, but may be found in getSupportInfo
    | 'css'
    | 'less'
    | 'scss'
    | 'typescript'
    | 'json'
    | 'json-stringify'
    | 'graphql'
    | 'markdown';

type TrailingCommaOption = 'none' | 'es5' | 'all';

interface PrettierSupportInfo {
    languages: {
        name: string;
        since: string;
        parsers: ParserOption[];
        group?: string;
        tmScope: string;
        aceMode: string;
        codemirrorMode: string;
        codemirrorMimeType: string;
        aliases?: string[];
        extensions: string[];
        filenames?: string[];
        linguistLanguageId: number;
        /**
         * Optional since prettier 1.14.0
         */
        vscodeLanguageIds?: string[];
    }[];
}

/**
 * Prettier configuration
 */
export interface PrettierConfig {
    printWidth: number;
    tabWidth: number;
    singleQuote: boolean;
    trailingComma: TrailingCommaOption;
    bracketSpacing: boolean;
    jsxBracketSameLine: boolean;
    parser: ParserOption;
    semi: boolean;
    useTabs: boolean;
    proseWrap: 'preserve' | 'always' | 'never';
    arrowParens: 'avoid' | 'always';
    rangeStart: number;
    rangeEnd: number;
    filepath: string;
    jsxSingleQuote: boolean;
    htmlWhitespaceSensitivity: 'css' | 'strict' | 'ignore';
    endOfLine: 'auto' | 'lf' | 'crlf' | 'cr';
    quoteProps: 'as-needed' | 'consistent' | 'preserve';
}
/**
 * prettier-vscode specific configuration
 */
interface ExtensionConfig {
    /**
     * Use 'prettier-eslint' instead of 'prettier'.
     * Other settings will only be fallbacks in case they could not be inferred from eslint rules.
     */
    eslintIntegration: boolean;
    /**
     * Use 'prettier-tslint' instead of 'prettier'.
     * Other settings will only be fallbacks in case they could not be inferred from tslint rules.
     */
    tslintIntegration: boolean;
    /**
     * Use 'prettier-stylelint' instead of 'prettier'.
     * Other settings will only be fallbacks in case they could not be inferred from eslint rules.
     */
    stylelintIntegration: boolean;
    /**
     * Path to '.prettierignore' or similar.
     */
    ignorePath: string;
    /**
     * Path to prettier config file.
     */
    config: string;
    /**
     * If true will skip formatting if a prettierconfig isn't found.
     */
    requireConfig: boolean;
    /**
     * Array of language IDs to ignore
     */
    disableLanguages: string[];
}

/**
 * Configuration for prettier-vscode
 */
export type PrettierVSCodeConfig = ExtensionConfig & PrettierConfig;
export interface Prettier {
    format: (text: string, options?: Partial<PrettierConfig>) => string;
    resolveConfig: (
        filePath: string,
        options?: {
            /**
             * Use cache, defaults to true.
             */
            useCache?: boolean;
            /**
             * read editorconfig, defaults to false.
             */
            editorconfig?: boolean;
            /**
             * Path to the config file to use, optional.
             */
            config?: string;
        }
    ) => Promise<PrettierConfig>;
    clearConfigCache: () => void;
    getSupportInfo(version?: string): PrettierSupportInfo;
    readonly version: string;
}